requirements.txt
 Django>=1.11,<2.0
 psycopg2
 djangorestframework